refactor: simplify ZoomableContainer by utilizing useZoomable hook
- Removed redundant state management and event handling from ZoomableContainer.
- Integrated useZoomable hook to encapsulate zoom and pan logic, enhancing code clarity.
- Updated type definitions for onWindowChange and renderContent props for better type safety.
- Improved overall structure and maintainability of the ZoomableContainer component.
